requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
dominate==2.6.0
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
numpy==1.20.3
Pillow==8.2.0
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
sklearn==0.0
threadpoolctl==2.1.0
urllib3==1.26.5
visitor==0.1.3
Werkzeug==2.0.1
